+++ src/datasets/domain/dtos/DatasetLicenseUpdateRequest.ts
@@ -1,0 +1,6 @@
+import { CustomTerms } from '../models/Dataset'
+
+export interface DatasetLicenseUpdateRequest {
+  name?: string
+  customTerms?: CustomTerms
+}