+++ src/datasets/index.ts
@@ -32,6 +32,8 @@ import { SetAvailableLicensesForDatasetType } from './domain/useCases/SetAvailab
 import { DeleteDatasetType } from './domain/useCases/DeleteDatasetType'
 import { GetDatasetCitationInOtherFormats } from './domain/useCases/GetDatasetCitationInOtherFormats'
 import { GetDatasetTemplates } from './domain/useCases/GetDatasetTemplates'
+import { UpdateTermsOfAccess } from './domain/useCases/UpdateTermsOfAccess'
+import { UpdateDatasetLicense } from './domain/useCases/UpdateDatasetLicense'
 
 const datasetsRepository = new DatasetsRepository()
 
@@ -80,6 +82,8 @@ const setAvailableLicensesForDatasetType = new SetAvailableLicensesForDatasetTyp
 const deleteDatasetType = new DeleteDatasetType(datasetsRepository)
 const getDatasetCitationInOtherFormats = new GetDatasetCitationInOtherFormats(datasetsRepository)
 const getDatasetTemplates = new GetDatasetTemplates(datasetsRepository)
+const updateTermsOfAccess = new UpdateTermsOfAccess(datasetsRepository)
+const updateDatasetLicense = new UpdateDatasetLicense(datasetsRepository)
 
 export {
   getDataset,
@@ -104,12 +108,14 @@ export {
   getDatasetAvailableCategories,
   getDatasetCitationInOtherFormats,
   getDatasetTemplates,
+  updateTermsOfAccess,
   getDatasetAvailableDatasetTypes,
   getDatasetAvailableDatasetType,
   addDatasetType,
   linkDatasetTypeWithMetadataBlocks,
   setAvailableLicensesForDatasetType,
-  deleteDatasetType
+  deleteDatasetType,
+  updateDatasetLicense
 }
 export { DatasetNotNumberedVersion } from './domain/models/DatasetNotNumberedVersion'
 export { DatasetUserPermissions } from './domain/models/DatasetUserPermissions'
@@ -136,12 +142,14 @@ export {
   DatasetMetadataBlockValuesDTO,
   DatasetMetadataChildFieldValueDTO
 } from './domain/dtos/DatasetDTO'
+export { DatasetLicenseUpdateRequest } from './domain/dtos/DatasetLicenseUpdateRequest'
 export { DatasetDeaccessionDTO } from './domain/dtos/DatasetDeaccessionDTO'
 export { CreatedDatasetIdentifiers } from './domain/models/CreatedDatasetIdentifiers'
 export { VersionUpdateType } from './domain/models/Dataset'
 export {
   DatasetVersionSummaryInfo,
-  DatasetVersionSummaryStringValues
+  DatasetVersionSummaryStringValues,
+  DatasetVersionSummarySubset
 } from './domain/models/DatasetVersionSummaryInfo'
 export { DatasetLinkedCollection } from './domain/models/DatasetLinkedCollection'
 export { DatasetType } from './domain/models/DatasetType'