+++ src/files/domain/repositories/IFilesRepository.ts
@@ -10,7 +10,7 @@ import { FileUploadDestination } from '../models/FileUploadDestination'
 import { UploadedFileDTO } from '../dtos/UploadedFileDTO'
 import { UpdateFileMetadataDTO } from '../dtos/UpdateFileMetadataDTO'
 import { RestrictFileDTO } from '../dtos/RestrictFileDTO'
-import { FileVersionSummaryInfo } from '../models/FileVersionSummaryInfo'
+import { FileVersionSummarySubset } from '../models/FileVersionSummaryInfo'
 
 export interface IFilesRepository {
   getDatasetFiles(
@@ -88,7 +88,11 @@ export interface IFilesRepository {
     replace?: boolean
   ): Promise<void>
 
-  getFileVersionSummaries(fileId: number | string): Promise<FileVersionSummaryInfo[]>
+  getFileVersionSummaries(
+    fileId: number | string,
+    limit?: number,
+    offset?: number
+  ): Promise<FileVersionSummarySubset>
 
   isFileDeleted(fileId: number | string): Promise<boolean>
 }